+++ coordinator/terraform/aws/applications/domainrecords/main.tf
@@ -16,7 +16,6 @@
 # Only needs to be used once
 
 terraform {
-  required_version = "~> 1.2.3"
   required_providers {
     aws = {
       source  = "hashicorp/aws"