+++ coordinator/terraform/aws/modules/keydb/main.tf
@@ -15,7 +15,6 @@
 data "aws_region" "current" {}
 
 terraform {
-  required_version = "~> 1.2.3"
   required_providers {
     aws = {
       source  = "hashicorp/aws"