+++ coordinator/terraform/aws/modules/publickeyhostingcloudfront/main.tf
@@ -13,7 +13,6 @@
 # limitations under the License.
 
 terraform {
-  required_version = "~> 1.2.3"
   required_providers {
     aws = {
       source = "hashicorp/aws"