+++ coordinator/terraform/aws/modules/sslcertificateprovider/main.tf
@@ -16,7 +16,6 @@ locals {
 }
 
 terraform {
-  required_version = "~> 1.2.3"
   required_providers {
     aws = {
       source  = "hashicorp/aws"