+++ coordinator/terraform/gcp/modules/keydb/files/02_create_expiry_time_index.sql
@@ -1,1 +1,0 @@
-CREATE INDEX KeySetsByExpiryTime ON KeySets(ExpiryTime)