+++ coordinator/terraform/gcp/modules/keydb/files/03_add_activation_time_column.sql
@@ -1,1 +1,0 @@
-ALTER TABLE KeySets ADD COLUMN ActivationTime TIMESTAMP