+++ coordinator/terraform/gcp/modules/keydb/files/04_create_activation_expiry_index.sql
@@ -1,1 +1,0 @@
-CREATE INDEX KeySetsByExpiryActivationDesc ON KeySets(ExpiryTime DESC, ActivationTime DESC)