+++ operator/terraform/aws/modules/bazel/terraform_version.tf
@@ -15,7 +15,6 @@
  */
 
 terraform {
-  required_version = "~> 1.2.3"
   required_providers {
     external = {
       source  = "hashicorp/external"