+++ operator/terraform/aws/modules/worker/main.tf
@@ -15,7 +15,6 @@
  */
 
 terraform {
-  required_version = "~> 1.2.3"
   required_providers {
     aws = {
       source  = "hashicorp/aws"